[#207] breadcrumb aren't shared components
Moving them in their respective places.
