+++ src/brokers/broker-pods/components/BrokerPodsBreadcrumb/BrokerPodsBreadcrumb.test.tsx
@@ -1,4 +1,4 @@
-import { screen, fireEvent, render, waitForI18n } from '../../test-utils';
+import { screen, fireEvent, render, waitForI18n } from '../../../../test-utils';
 import { BrokerPodsBreadcrumb } from './BrokerPodsBreadcrumb';
 import { MemoryRouter } from 'react-router-dom';
