+++ src/shared-components/index.ts
@@ -1,2 +1,1 @@
 export * from './Loading';
-export * from './BrokerPodsBreadcrumb';
